+++ packages/docs/src/stories/Heading.stories.tsx
@@ -6,7 +6,7 @@ export default {
   title: 'Typography/Heading',
   component: Heading,
   args: {
-    sizeH: "lg",
+    sizeHead: "xl",
     type: "h2"
   },
   argTypes: {
@@ -14,8 +14,8 @@ export default {
       options: ["h1","h2","h3","h4","h5","h6"],
       control:{ type: 'inline-radio'}
     },
-    sizeH: {
-      options: ["lg","2xl","4xl","6xl","8xl","9xl"],
+    sizeHead: {
+      options: ["xl","2xl","4xl","6xl","8xl","9xl"],
       control:{ type: 'inline-radio'}
     },
   }
@@ -24,11 +24,13 @@ export default {
 export const Primary: StoryObj<HeadingProps> = {
   args: {
     children: 'Title H2',
+
+
   },
 }
 export const DinamicTag: StoryObj<HeadingProps> = {
     args:{
-        sizeH: "8xl", 
+        sizeHead: "8xl", 
         type: 'h1',
         children: 'Title H1',	
     }