+++ packages/react/src/components/CheckBox.tsx
@@ -8,12 +8,15 @@ export function CheckBox(props: CheckBoxProps) {
   return (
     <CheckBoxComp.Root
       {...props}
-      className={`box-border flex h-6 w-6 cursor-pointer items-center justify-center overflow-hidden 
-    rounded-Myll_xs border-2 border-solid border-Myll_gray900 bg-Myll_gray900 leading-none transition-all delay-150 duration-700 
-    ease-in-out focus:border-Myll_ignite300 data-[state=checked]:translate-y-0 data-[state=checked]:border-Myll_ignite300 data-[state=checked]:bg-Myll_ignite300  
+      className={`box-border flex h-6 w-6 cursor-pointer items-center justify-center 
+    overflow-hidden rounded-Myll_xs border-2 border-solid border-Myll_gray900 bg-Myll_gray900 leading-none 
+ focus:border-Myll_ignite300  data-[state=checked]:border-Myll_ignite300 data-[state=checked]:bg-Myll_ignite300  
      `}
     >
-      <CheckBoxComp.Indicator asChild className={`h-4 w-4 text-Myll_white `}>
+      <CheckBoxComp.Indicator
+        asChild
+        className={`h-4 w-4 text-Myll_white data-[state=checked]:animate-slideIn data-[state=unchecked]:animate-swipeOut `}
+      >
         <Check weight="bold" />
       </CheckBoxComp.Indicator>
     </CheckBoxComp.Root>