+++ packages/eslint-config/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @mll-ui/eslint-config
 
+## 3.0.0
+
+### Major Changes
+
+- Install and update Tailwind css Variants
+
 ## 2.0.0
 
 ### Major Changes